+++ modules/home-manager.nix
@@ -1,0 +1,11 @@
+{ config, pkgs, ... }:
+
+let
+  home-manager = builtins.fetchTarball
+    "https://github.com/nix-community/home-manager/archive/release-23.11.tar.gz";
+in
+{
+  imports = [
+    (import "${home-manager}/nixos")
+  ];
+}