+++ users/jwinn/dotfiles/config/bash/env.bash
@@ -1,0 +1,7 @@
+# shellcheck shell=bash
+
+# WARNING: this will $BASH_ENV to get a correct startup sequence
+
+# shellcheck source-path=SCRIPTDIR
+. "/home/jwinn/.config/shell/env.sh"
+