+++ users/jwinn/dotfiles/config/zsh/logout.zsh
@@ -1,0 +1,5 @@
+# shellcheck shell=bash
+
+# shellcheck source-path=SCRIPTDIR
+. "/home/jwinn/.config/shell/logout.sh"
+