+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog],
 and this project adheres to [Semantic Versioning].
 
+## [unreleased]
+
+### Added
+
+-   Home Manager (HM) module-based NixOS configuration
+
 ## [0.1.0] - 2024-02-18
 
 ### Added